fix(leaderboard): Faulty logic in ScoringCriteria queries

+++ api/internal/lib/models/leaderboard.go
@@ -5,7 +5,7 @@ type ScoringInput struct {
 	PlayerID            PlayerID
 	Name                string
 	NumScores           int64
-	TotalPoints         int32
+	TotalPoints         int64
 	PointsFromPenalties int64
 	PointsFromBonuses   int64
 }
